Continued debugging of request message body
Another possibility is that the message parser is reading lines faster than they are being sent by the test client. This woudl explain why the request message always states the correct content-length, but only sometimes fails to be read. This would also explain why adding a simple Thread.sleep(1) to the parser seems to remove the issue.

What we need is a solution that sits and waits for body content, if, and only if, we are sure there is more content to be read. Otherwise, we end up waiting for a line that never comes.