Implemented basic multithreading.
Most of the business logic in HTTPServer has been refactored into an HTTPHandler class.
The HTTPHandler class implements Runnable, and the HTTPServer class now instantiates this Runnable on a new thread for each connection request on the server socket.
The ServerTestClient class has also had multithreading implemented, along with a class that creates the desired number of Threads running the test class.
All tests seem to still pass (I really need to rewrite that thing to be automated...), and the server isn't throwing any exceptions (though the PokerSend class did have issues with concurrency access conflicts, which have been fixed.).